mfcc_scripts/requirements.txt
numpy>=1.20.1
Jinja2>=3.0.1
scipy>=1.7.3
PyYAML>=6.0
